requirements.txt
PyQt6==6.6.1
httpx==0.27.0
python-dotenv==1.0.1
